+++ src/frontend/apps/lti_site/apps/.cookiecutter/{{cookiecutter.app_name}}/components/Routes.tsx
@@ -6,8 +6,13 @@ import { Loader } from 'lib-components';
 import {
   ErrorComponentsProps,
   FullScreenError,
+  FULL_SCREEN_ERROR_ROUTE,
+  useAppConfig,
 } from 'lib-components';
-import { FULL_SCREEN_ERROR_ROUTE } from 'lib-components';
+
+import { PortabilityRequest } from 'components/PortabilityRequest';
+import { RESOURCE_PORTABILITY_REQUEST_ROUTE } from 'components/PortabilityRequest/route';
+
 import { DASHBOARD_{{ cookiecutter.app_name|upper }}_ROUTE } from './Dashboard{{cookiecutter.model}}/route';
 import { RedirectOnLoad } from './RedirectOnLoad';
 import { REDIRECT_ON_LOAD_ROUTE } from './RedirectOnLoad/route';
@@ -15,6 +20,8 @@ import { REDIRECT_ON_LOAD_ROUTE } from './RedirectOnLoad/route';
 const Dashboard{{cookiecutter.model}} = lazy(() => import('./Dashboard{{cookiecutter.model}}'));
 
 export const Routes = () => {
+  const appData = useAppConfig();
+
   return (
     <Suspense fallback={<Loader />}>
       <MemoryRouter>
@@ -30,6 +37,16 @@ export const Routes = () => {
           render={({ match }) => (
             <FullScreenError
               code={match.params.code as ErrorComponentsProps['code']}
+            />
+          )}
+        />
+
+        <Route
+          exact
+          path={RESOURCE_PORTABILITY_REQUEST_ROUTE()}
+          render={() => (
+            <PortabilityRequest
+              portability={appData.portability!}
             />
           )}
         />