+++ src/frontend/apps/lti_site/apps/classroom/components/RedirectOnLoad/index.spec.tsx
@@ -5,6 +5,8 @@ import {
   useAppConfig,
   appState,
 } from 'lib-components';
+
+import { RESOURCE_PORTABILITY_REQUEST_ROUTE } from 'components/PortabilityRequest/route';
 import render from 'utils/tests/render';
 
 import { RedirectOnLoad } from '.';
@@ -83,5 +85,25 @@ describe('<RedirectOnLoad />', () => {
     });
 
     getByText('Dashboard');
+  });
+
+  it('redirects to portability if app state requires it', async () => {
+    mockedUseAppConfig.mockReturnValue({
+      flags: { classroom: true },
+      state: appState.PORTABILITY,
+    } as any);
+
+    const { getByText } = render(<RedirectOnLoad />, {
+      routerOptions: {
+        routes: [
+          {
+            path: RESOURCE_PORTABILITY_REQUEST_ROUTE(),
+            render: () => <span>Portability request</span>,
+          },
+        ],
+      },
+    });
+
+    getByText('Portability request');
   });
 });